final-project p2: рефакторинг

+++ ewm-main/src/main/java/ru/practicum/service/impl/CategoriesServiceImpl.java
@@ -12,6 +12,7 @@ import ru.practicum.model.dto.category.CategoryDto;
 import ru.practicum.model.dto.category.NewCategoryDto;
 import ru.practicum.model.entity.Category;
 import ru.practicum.repository.CategoryRepository;
+import ru.practicum.repository.event.EventRepository;
 import ru.practicum.service.CategoriesService;
 
 import java.util.List;
@@ -22,21 +23,25 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class CategoriesServiceImpl implements CategoriesService {
     private final CategoryRepository categoryRepository;
+    private final EventRepository eventRepository;
     private final CategoryMapper categoryMapper;
+
     @Override
     @Transactional
     public CategoryDto addCategory(NewCategoryDto newCategoryDto) {
-        if(categoryRepository.existsByName(newCategoryDto.getName()))
+        if (categoryRepository.existsByName(newCategoryDto.getName()))
             throw new ConflictException("категория уже существует");
-
         return categoryMapper.toDto(categoryRepository.save(categoryMapper.toEntity(newCategoryDto)));
     }
 
     @Override
     @Transactional
     public void deleteCategory(Long catId) {
-        categoryRepository.findById(catId)
+        Category category = categoryRepository.findById(catId)
                 .orElseThrow(() -> new NotFoundException("category is not found"));
+        if (eventRepository.existsByCategoryId(category.getId())) {
+            throw new ConflictException("Невозможно удалить категорию, так как она связана с событием(событиями)");
+        }
         categoryRepository.deleteById(catId);
     }
 
@@ -45,8 +50,9 @@ public class CategoriesServiceImpl implements CategoriesService {
     public CategoryDto updateCategory(Long catId, CategoryDto categoryDto) {
         Category category = categoryRepository.findById(catId)
                 .orElseThrow(() -> new NotFoundException("category is not found"));
-        if(category.getName().equals(categoryDto.getName()))
-            throw new ConflictException("Такая категория уже существует");
+        if (!categoryDto.getName().equals(category.getName()) && categoryRepository.existsByName(categoryDto.getName())) {
+            throw new ConflictException("Имя уже используется другой категорией");
+        }
         category.setName(categoryDto.getName());
         return categoryMapper.toDto(category);
     }
